add helper functions for handler
