+++ xray/handler.go
@@ -9,6 +9,7 @@
 package xray
 
 import (
+	"bytes"
 	"context"
 	"net"
 	"net/http"
@@ -116,7 +117,8 @@ func Handler(sn SegmentNamer, h http.Handler) http.Handler {
 }
 
 func httpTrace(seg *Segment, h http.Handler, w http.ResponseWriter, r *http.Request, traceHeader *header.Header) {
-	traceID := seg.TraceHeaderID(traceHeader)
+	httpCaptureRequest(seg, r)
+	traceID := traceHeaderID(seg, traceHeader)
 	w.Header().Set(TraceIDHeaderKey, traceID)
 
 	capturer := &responseCapturer{w, 200, 0}
@@ -125,7 +127,7 @@ func httpTrace(seg *Segment, h http.Handler, w http.ResponseWriter, r *http.Requ
 
 	cLen, _ := strconv.Atoi(capturer.Header().Get("Content-Length"))
 	seg.GetHTTP().GetResponse().ContentLength = cLen
-	seg.HTTPCapture(capturer.status)
+	httpCaptureResponse(seg, capturer.status)
 }
 
 func clientIP(r *http.Request) (string, bool) {
@@ -145,4 +147,55 @@ func btoi(b bool) int {
 		return 1
 	}
 	return 0
+}
+
+// traceHeaderID return trace id for http header
+func traceHeaderID(seg *Segment, traceHeader *header.Header) string {
+	seg.Lock()
+	defer seg.Unlock()
+
+	var respHeader bytes.Buffer
+	respHeader.WriteString("Root=")
+	respHeader.WriteString(seg.TraceID)
+
+	if traceHeader.SamplingDecision == header.Requested {
+		respHeader.WriteString(";Sampled=")
+		respHeader.WriteString(strconv.Itoa(btoi(seg.Sampled)))
+	}
+
+	return respHeader.String()
+}
+
+// httpCaptureResponse fill response by http status code
+func httpCaptureResponse(seg *Segment, statusCode int) {
+	seg.Lock()
+	defer seg.Unlock()
+
+	seg.GetHTTP().GetResponse().Status = statusCode
+
+	if statusCode >= 400 && statusCode < 500 {
+		seg.Error = true
+	}
+	if statusCode == 429 {
+		seg.Throttle = true
+	}
+	if statusCode >= 500 && statusCode < 600 {
+		seg.Fault = true
+	}
+}
+
+// httpCaptureRequest fill request data by http.Request
+func httpCaptureRequest(seg *Segment, r *http.Request) {
+	seg.Lock()
+	defer seg.Unlock()
+
+	scheme := "https://"
+	if r.TLS == nil {
+		scheme = "http://"
+	}
+
+	seg.GetHTTP().GetRequest().Method = r.Method
+	seg.GetHTTP().GetRequest().URL = scheme + r.Host + r.URL.Path
+	seg.GetHTTP().GetRequest().ClientIP, seg.GetHTTP().GetRequest().XForwardedFor = clientIP(r)
+	seg.GetHTTP().GetRequest().UserAgent = r.UserAgent()
 }
